:recycle: Move deployment API import tests to dedicated subfolder
