+++ test/4_deployment_api/import/import_from_xml_local.js
@@ -1,11 +1,10 @@
 'use strict';
 
 const should = require('should');
-const path = require('path');
 
-const TestFixtureProvider = require('../../dist/commonjs/test_setup').TestFixtureProvider;
+const TestFixtureProvider = require('../../../dist/commonjs/test_setup').TestFixtureProvider;
 
-describe('Deployment API -> importBpmnFromFile', () => {
+describe('Deployment API -> importBpmnFromXml', () => {
 
   let testFixtureProvider;
   let defaultIdentity;
@@ -16,10 +15,10 @@ describe('Deployment API -> importBpmnFromFile', () => {
   const processModelIdNameMismatch = 'process_model_name_mismatch';
   const processModelIdTooManyProcesses = 'process_model_too_many_processes';
 
-  let processModelPath;
-  let processModelPathNoLanes;
-  let processModelPathNameMismatch;
-  let processModelPathTooManyProcesses;
+  let processModelAsXml;
+  let processModelNoLanesAsXml;
+  let processModelPathNameMismatchAsXml;
+  let processModelPathTooManyProcessesAsXml;
 
   before(async () => {
     testFixtureProvider = new TestFixtureProvider();
@@ -28,12 +27,10 @@ describe('Deployment API -> importBpmnFromFile', () => {
     defaultIdentity = testFixtureProvider.identities.defaultUser;
     restrictedIdentity = testFixtureProvider.identities.restrictedUser;
 
-    const bpmnFolderLocation = testFixtureProvider.getBpmnDirectoryPath();
-
-    processModelPath = path.join(bpmnFolderLocation, `${processModelId}.bpmn`);
-    processModelPathNoLanes = path.join(bpmnFolderLocation, `${processModelIdNoLanes}.bpmn`);
-    processModelPathNameMismatch = path.join(bpmnFolderLocation, `${processModelIdNameMismatch}.bpmn`);
-    processModelPathTooManyProcesses = path.join(bpmnFolderLocation, `${processModelIdTooManyProcesses}.bpmn`);
+    processModelAsXml = testFixtureProvider.readProcessModelFile(processModelId);
+    processModelNoLanesAsXml = testFixtureProvider.readProcessModelFile(processModelIdNoLanes);
+    processModelPathNameMismatchAsXml = testFixtureProvider.readProcessModelFile(processModelIdNameMismatch);
+    processModelPathTooManyProcessesAsXml = testFixtureProvider.readProcessModelFile(processModelIdTooManyProcesses);
   });
 
   after(async () => {
@@ -41,40 +38,62 @@ describe('Deployment API -> importBpmnFromFile', () => {
   });
 
   it('should successfully import a ProcessModel', async () => {
-    await testFixtureProvider.deploymentApiService.importBpmnFromFile(defaultIdentity, processModelPath, processModelId, true);
+
+    const importPayload = {
+      name: processModelId,
+      xml: processModelAsXml,
+      overwriteExisting: true,
+    };
+
+    await testFixtureProvider.deploymentApiService.importBpmnFromXml(defaultIdentity, importPayload);
     await assertThatImportWasSuccessful();
   });
 
   it('should successfully import a ProcessModel without any lanes', async () => {
 
-    const overwriteIfExists = true;
-    await testFixtureProvider
-      .deploymentApiService
-      .importBpmnFromFile(defaultIdentity, processModelPathNoLanes, processModelIdNoLanes, overwriteIfExists);
+    const importPayload = {
+      name: processModelIdNoLanes,
+      xml: processModelNoLanesAsXml,
+      overwriteExisting: true,
+    };
 
+    await testFixtureProvider.deploymentApiService.importBpmnFromXml(defaultIdentity, importPayload);
     await assertThatImportWasSuccessful();
   });
 
   it('should fail to import a ProcessModel, if a process model by the same name exists, and overwriteExisting is set to false', async () => {
 
     try {
-      // Run this twice to ensure that this test case is always executable.
-      await testFixtureProvider.deploymentApiService.importBpmnFromFile(defaultIdentity, processModelPath, processModelId, false);
-      await testFixtureProvider.deploymentApiService.importBpmnFromFile(defaultIdentity, processModelPath, processModelId, false);
+      const importPayload = {
+        name: processModelId,
+        xml: processModelAsXml,
+        overwriteExisting: false,
+      };
 
-      should.fail(undefined, 'error', 'This request should have failed, because a ProcessModel already exists!');
+      // Run this twice to ensure that this test case is always executable.
+      await testFixtureProvider.deploymentApiService.importBpmnFromXml(defaultIdentity, importPayload);
+      await testFixtureProvider.deploymentApiService.importBpmnFromXml(defaultIdentity, importPayload);
+
+      should.fail(undefined, 'error', 'This request should have failed, because the process model already exists!');
     } catch (error) {
       const expectedErrorCode = 409;
       const expectedErrorMessage = /already exists/i;
       should(error.code).be.eql(expectedErrorCode);
       should(error.message).be.match(expectedErrorMessage);
     }
+
   });
 
   it('should fail to import a ProcessModel, when the user is not authenticated', async () => {
 
+    const importPayload = {
+      name: processModelId,
+      xml: processModelAsXml,
+      overwriteExisting: false,
+    };
+
     try {
-      await testFixtureProvider.deploymentApiService.importBpmnFromFile(undefined, processModelPath);
+      await testFixtureProvider.deploymentApiService.importBpmnFromXml(undefined, importPayload);
       should.fail({}, 'error', 'This request should have failed, due to missing user authentication!');
     } catch (error) {
       const expectedErrorCode = 401;
@@ -86,8 +105,14 @@ describe('Deployment API -> importBpmnFromFile', () => {
 
   it('should fail to import a ProcessModel, when the user is forbidden to see the process instance result', async () => {
 
+    const importPayload = {
+      name: processModelId,
+      xml: processModelAsXml,
+      overwriteExisting: false,
+    };
+
     try {
-      await testFixtureProvider.deploymentApiService.importBpmnFromFile(restrictedIdentity, processModelPath);
+      await testFixtureProvider.deploymentApiService.importBpmnFromXml(restrictedIdentity, importPayload);
       should.fail(undefined, 'error', 'This request should have failed, due to a missing claim!');
     } catch (error) {
       const expectedErrorCode = 403;
@@ -101,11 +126,14 @@ describe('Deployment API -> importBpmnFromFile', () => {
   // Otherwise the ProcessModel would not be retrievable.
   it('should fail to import a ProcessModel, when the ProcessModel name does not match the ProcessDefinition name', async () => {
 
-    try {
-      await testFixtureProvider
-        .deploymentApiService
-        .importBpmnFromFile(defaultIdentity, processModelPathNameMismatch, processModelIdNameMismatch);
+    const importPayload = {
+      name: processModelIdNameMismatch,
+      xml: processModelPathNameMismatchAsXml,
+      overwriteExisting: false,
+    };
 
+    try {
+      await testFixtureProvider.deploymentApiService.importBpmnFromXml(defaultIdentity, importPayload);
       should.fail(undefined, 'error', 'This request should have failed, because ProcessModel name differs from the ProcessDefinitions name!');
     } catch (error) {
       const expectedErrorMessage = /ProcessModel contained within the diagram.*?must also use the name/i;
@@ -120,11 +148,14 @@ describe('Deployment API -> importBpmnFromFile', () => {
   // Otherwise, BPMN Studio will be unable to use the Runtime at all.
   it('should fail to import a ProcessModel, when the file contains more than one Process', async () => {
 
-    try {
-      await testFixtureProvider
-        .deploymentApiService
-        .importBpmnFromFile(defaultIdentity, processModelPathTooManyProcesses, processModelIdTooManyProcesses);
+    const importPayload = {
+      name: processModelIdTooManyProcesses,
+      xml: processModelPathTooManyProcessesAsXml,
+      overwriteExisting: false,
+    };
 
+    try {
+      await testFixtureProvider.deploymentApiService.importBpmnFromXml(defaultIdentity, importPayload);
       should.fail(undefined, 'error', 'This request should have failed, because the ProcessDefinition has more than one model!');
     } catch (error) {
       const expectedErrorMessage = /contains more than one ProcessModel/i;
